backend/requirements.txt
accelerate==1.13.0
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.8.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.13.0
APScheduler==3.11.2
async-timeout==5.0.1
asyncua==1.1.8
attrs==26.1.0
av==14.2.0
backoff==2.2.1
bitsandbytes==0.47.0
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.115.14
filelock==3.17.0
Flask==3.1.3
flask-cors==6.0.2
flask-sock==0.7.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
google-crc32c==1.8.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.3
inference==1.2.3
inference-models==0.24.4
inference-sdk==1.2.3
iniconfig==2.3.0
iopath==0.1.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
kiwisolver==1.5.0
langdetect==1.0.9
lazy-loader==0.5
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
num2words==0.5.14
numpy==2.2.6
nvidia-ml-py==12.575.51
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.13.0.92
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
peft==0.19.1
pillow==11.3.0
platformdirs==4.9.6
pluggy==1.6.0
polars==1.40.0
polars-runtime-32==1.40.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.25.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pybase64==1.0.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
pyDeprecate==0.5.0
pydot==2.0.0
pyee==13.0.1
Pygments==2.20.0
PyJWT==2.12.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdfium2==5.7.0
pyserial==3.5
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-doctr==1.0.1
python-dotenv==1.0.1
pytz==2026.1.post1
pyvips==2.2.3
PyYAML==6.0.3
qrcode==8.0
RapidFuzz==3.14.5
redis==5.0.8
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rf-clip==1.1
RF-SAM-2==1.0.3
rf-segment-anything==1.0
rf_groundingdino==0.3.0
rich==14.3.4
s3transfer==0.15.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
segmentation_models_pytorch==0.5.0
sentencepiece==0.2.1
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
simple-websocket==1.1.0
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
structlog==24.4.0
supervision==0.27.0.post2
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.26
tldextract==5.1.3
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
trackers==2.3.0
transformers==5.2.0
twilio==9.3.8
typer==0.16.0
typer-slim==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ultralytics==8.4.39
ultralytics-thop==2.0.18
urllib3==2.6.3
validators==0.35.0
wait_for2==0.3.2
wcwidth==0.6.0
Werkzeug==3.1.8
wsproto==1.3.2
yapf==0.43.0
yarl==1.23.0
zeep==4.3.2
zxing-cpp==2.2.0
